Play: fix draw bug so delta applied correctly
In a draw, if the delta is non-zero, the higher-rated player should lose
points and the lower rated player should gain points. Since the
assignment of w/l is irrelevant in the case of a draw, Play should
handle this special case by applying the rating gains and losses to the
appropriate players.

Includes tests for this case and others, keeping 100% test coverage.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>